Update to QT6

+++ src/external/qt-color-widgets/src/color_utils.hpp
@@ -37,7 +37,7 @@ inline qreal color_lumaF(const QColor& c)
 {
     return 0.30 * c.redF() + 0.59 * c.greenF() + 0.11 * c.blueF();
 }
-QColor color_from_lch(qreal hue, qreal chroma, qreal luma, qreal alpha = 1 );
+QColor color_from_lch(float hue, float chroma, float luma, float alpha = 1 );
 
 inline QColor rainbow_lch(qreal hue)
 {
@@ -64,7 +64,7 @@ inline qreal color_HSL_saturationF(const QColor& col)
     return c / (1-qAbs(2*l-1));
 }
 
-QColor color_from_hsl(qreal hue, qreal sat, qreal lig, qreal alpha = 1 );
+QColor color_from_hsl(float hue, float sat, float lig, float alpha = 1 );
 
 } // namespace detail
 } // namespace color_widgets
