+++ src/external/qt-color-widgets/src/color_wheel.cpp
@@ -55,7 +55,7 @@ public:
     QPixmap hue_ring;
     QImage inner_selector;
     DisplayFlags display_flags;
-    QColor (*color_from)(qreal,qreal,qreal,qreal);
+    QColor (*color_from)(float,float,float,float);
     QColor (*rainbow_from_hue)(qreal);
     int max_size = 128;
